refactoring auth modals
eliminated user-details from store
