+++ react-app/src/components/SplashPage/index.js
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useHistory } from "react-router-dom";
 import { getAllSongs } from "../../store/song";
-import LoginForm from "../auth/LoginForm";
-import SignUpForm from "../auth/SignUpForm";
+import LoginForm from "../../modals/LoginForm";
+import SignUpForm from "../../modals/SignUpForm";
 import { Modal } from "../Context/Modal";
 import Logo from "../Icons/Logo";
 import ShowcaseSongs from "../Modules/ShowcaseSongs";
