+++ react-app/src/store/index.js
@@ -3,7 +3,6 @@ import thunk from "redux-thunk";
 import session from "./session";
 import songs from "./song";
 import player from "./player";
-import details from "./user-details";
 import playlists from "./playlist";
 import users from "./user";
 
@@ -11,7 +10,6 @@ const rootReducer = combineReducers({
   session,
   songs,
   player,
-  details,
   playlists,
   users,
 });
